+++ Lab1/NodeTree.h
@@ -1,0 +1,14 @@
+#pragma once
+
+template <typename T>
+class NodeTree {
+public:
+	T data;
+	NodeTree<T>* parent;
+	vector<NodeTree<T>*> children;
+	vector<int> weight;
+
+	NodeTree();
+	NodeTree(T data);
+	NodeTree(T data, NodeTree<T>* parent);
+};
